restructuring of packages, adding logout+proper user authentication using flask-login
