+++ foodtalk/templates/layout.html
@@ -27,9 +27,13 @@
                   <a class="nav-item nav-link" href="/">Left</a>
               </div>
               <div class="navbar-nav">
-                <a class="nav-item nav-link" href="/login">Login</a>
+                {% if current_user.is_authenticated %}
+                <a class="nav-item nav-link" href="/logout">Login</a>
+                {% else %}
+                <a class="nav-item nav-link" href="/login">Logout</a>
                 <a class="nav-item nav-link" href="/register">Register</a>
                 <a class="nav-item nav-link" href="/register_business">Business Registration</a>
+                {% endif %}
               </div>
             </div>
           </div>
